git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@2135 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/gameseg.cpp
@@ -1986,12 +1986,12 @@ void ComputeVertexNormals (void)
 
 for (i = gameData.segs.nVertices, pp = gameData.segs.points; i; i--, pp++) {
 	if (1 < (h = pp->p3_normal.nFaces)) {
-		pp->p3_normal.vNormal /= h;
+		pp->p3_normal.vNormal /= (float) h;
 		/*
 		pp->p3_normal.vNormal[Y] /= h;
 		pp->p3_normal.vNormal[Z] /= h;
 		*/
-		fVector::Normalize(pp->p3_normal.vNormal);
+		fVector::Normalize (pp->p3_normal.vNormal);
 		}
 	pp->p3_normal.nFaces = 1;
 	}
